+++ core/src/builtin_widgets/align.rs
@@ -68,13 +68,13 @@ pub struct VAlignWidget {
 impl Declare for HAlignWidget {
   type Builder = FatObj<()>;
   #[inline]
-  fn declare_builder() -> Self::Builder { FatObj::new(()) }
+  fn declarer() -> Self::Builder { FatObj::new(()) }
 }
 
 impl Declare for VAlignWidget {
   type Builder = FatObj<()>;
   #[inline]
-  fn declare_builder() -> Self::Builder { FatObj::new(()) }
+  fn declarer() -> Self::Builder { FatObj::new(()) }
 }
 
 impl Render for HAlignWidget {