+++ core/src/builtin_widgets/cursor.rs
@@ -11,7 +11,7 @@ pub struct Cursor {
 impl Declare for Cursor {
   type Builder = FatObj<()>;
   #[inline]
-  fn declare_builder() -> Self::Builder { FatObj::new(()) }
+  fn declarer() -> Self::Builder { FatObj::new(()) }
 }
 
 impl ComposeChild for Cursor {