+++ core/src/builtin_widgets/focus_node.rs
@@ -8,7 +8,7 @@ pub struct RequestFocus {
 impl Declare for RequestFocus {
   type Builder = FatObj<()>;
   #[inline]
-  fn declare_builder() -> Self::Builder { FatObj::new(()) }
+  fn declarer() -> Self::Builder { FatObj::new(()) }
 }
 
 impl ComposeChild for RequestFocus {