+++ core/src/builtin_widgets/scrollable.rs
@@ -26,7 +26,7 @@ pub struct ScrollableWidget {
 impl Declare for ScrollableWidget {
   type Builder = FatObj<()>;
   #[inline]
-  fn declare_builder() -> Self::Builder { FatObj::new(()) }
+  fn declarer() -> Self::Builder { FatObj::new(()) }
 }
 
 impl ComposeChild for ScrollableWidget {